store base components refactor
