+++ core/ui/src/commonMain/kotlin/com/stslex/core/ui/mvi/Store.kt
@@ -3,8 +3,24 @@ package com.stslex.core.ui.mvi
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.runtime.Stable
 import com.stslex.core.ui.components.SnackbarType
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
 
-interface Store {
+interface Store<S : Store.State, E : Store.Event, A : Store.Action> {
+
+    /** Flow of the state of the screen. */
+    val state: StateFlow<S>
+
+    /** Flow of events that are sent to the screen. */
+    val event: SharedFlow<E>
+
+    /**
+     * Sends an action to the store. Checks if the action is not the same as the last action.
+     * If the action is not the same as the last action, the last action is updated.
+     * The action is then processed.
+     * @param action - action to be sent
+     */
+    fun sendAction(action: A)
 
     interface State
