+++ feature/match/src/commonMain/kotlin/com/stslex/feature/match/navigation/MatchRouter.kt
@@ -1,6 +1,6 @@
 package com.stslex.feature.match.navigation
 
 import com.stslex.core.ui.mvi.Router
-import com.stslex.feature.match.ui.store.MatchStoreComponent.Navigation
+import com.stslex.feature.match.ui.store.MatchStore.Navigation
 
 interface MatchRouter : Router<Navigation>
